xen/mm: Update the limit for max_pages from UINT_MAX to ULONG_MAX
xen/arch/x86/dom0_build.c: Update the upper limit for Dom0's max_pages
in dom0_compute_nr_pages() to limit from UINT_MAX to ULONG_MAX.

tools/libs/stat/xenstat.c: Update the "no maximum" sentinel for max_pages
in xenstat_get_node() from UINT_MAX to ULONG_MAX, which is converted to
(unsigned long long)-1.

+++ xen/arch/x86/dom0_build.c
@@ -440,7 +440,7 @@ unsigned long __init dom0_compute_nr_pages(
         }
     }
 
-    d->max_pages = min_t(unsigned long, max_pages, UINT_MAX);
+    d->max_pages = min_t(unsigned long, max_pages, ULONG_MAX);
 
     return nr_pages;
 }
